refactor: create func handlerUser handle service user

+++ main.py
@@ -10,33 +10,29 @@ import service.user as userService
 app = Flask(__name__)
 
 # Method GET return ค่าข้อมูลของ user
-@app.route('/user', methods=['GET'])
-def getUser():
-    id = request.args.get('id')
-    return userService.getUsers(id)
+@app.route('/user', methods=['GET', 'POST', 'PUT', 'DELETE'])
+def handler_user():
+    if request.method == 'GET':
+        id = request.args.get('id')
+        return userService.getUsers(id)
     
-# Method POST เพิ่มข้อมูล user
-@app.route('/user', methods=['POST'])
-def createUser():
-    body = request.get_json()
-    return userService.createUser(body)
+    if request.method == 'POST':
+        body = request.get_json()
+        return userService.createUser(body)
 
-# Method PUT แก้ไขข้อมูล user ที่ id ตรงกับที่กำหนด
-@app.route('/user', methods=['PUT'])
-def editUser():
-    body = request.get_json()
-    id = request.args.get('id')
-    return userService.editUser(body, id)
+    if request.method == 'PUT':
+        body = request.get_json()
+        id = request.args.get('id')
+        return userService.editUser(body, id)
 
-# Method DELETE ลบข้อมูล user ที่ id ตรงตามที่กำหนด
-@app.route('/user', methods=['DELETE'])
-def deleteUser():
-    id = request.args.get('id')
-    return userService.deleteUser(id)
+    if request.method == 'DELETE':
+        id = request.args.get('id')
+        return userService.deleteUser(id)
+
 
 ### Cal age average ###
 @app.route('/user/avg', methods=['GET'])
-def calAvg():
+def handler_calculate_average_age():
    return userService.calAvg()
